backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "nbajinni-backend"
version = "0.1.0"
description = "NBAJinni FastAPI backend API"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
mangum = "^0.17.0"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
python-multipart = "^0.0.9"
python-dotenv = "^1.0.0"
slowapi = "^0.1.9"
boto3 = "^1.34.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
nbajinni-shared = { path = "../shared", develop = true }
structlog = "^25.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
black = "^24.4.0"
flake8 = "^7.0.0"
isort = "^5.13.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]